two-step: Add bgrt rotation quirk handling
The Ayaneo 3 has a bgrt image that is not rotated in respect to the
portrait screen. However, it lists a portrait bgrt orientation. This
causes us to skip applying the rotation to the bgrt image, making it
appear sideways. Therefore, before skipping the rotation, check if the
image is aligned height-wise on the portrait axis (i.e., width wise on
the landscape axis) and only apply the rotation if it is not.